feat: send ical events if updated at section update
